Move api code to its own namespace
