+++ custom_components/gree_custom/aiogree/__init__.py
@@ -1,0 +1,1 @@
+"""aiogree provides an interface to comunicate with a Gree device."""
