+++ custom_components/gree_custom/aiogree/api.py
@@ -14,24 +14,15 @@ import asyncio_dgram
 from attr import dataclass
 from Crypto.Cipher import AES
 
+from .const import (
+    DEFAULT_DEVICE_PORT,
+    GCM_ADD,
+    GCM_IV,
+    GREE_GENERIC_DEVICE_KEY,
+    GREE_GENERIC_DEVICE_KEY_GCM,
+)
+
 _LOGGER = logging.getLogger(__name__)
-
-GCM_IV = b"\x54\x40\x78\x44\x49\x67\x5a\x51\x6c\x5e\x63\x13"
-GCM_ADD = b"qualcomm-test"
-
-GREE_GENERIC_DEVICE_KEY = "a3K8Bx%2r8Y7#xDh"
-GREE_GENERIC_DEVICE_KEY_GCM = "{yxAHAY_Lm6pbC/<"
-
-MIN_TEMP_C = 16
-MAX_TEMP_C = 30
-
-MIN_TEMP_F = 61
-MAX_TEMP_F = 86
-
-DEFAULT_DEVICE_UID = 0
-DEFAULT_DEVICE_PORT = 7000
-DEFAULT_CONNECTION_MAX_ATTEMPTS = 5
-DEFAULT_CONNECTION_TIMEOUT = 10
 
 
 class GreeProp(Enum):
