+++ custom_components/gree_custom/select.py
@@ -14,6 +14,8 @@ from homeassistant.helpers.entity_platform import AddEntitiesCallback
 from homeassistant.helpers.restore_state import RestoreEntity
 from homeassistant.helpers.typing import UNDEFINED
 
+from .aiogree.api import GreeProp, TemperatureUnits
+from .aiogree.device import GreeDevice
 from .const import (
     CONF_DEVICES,
     CONF_DISABLE_AVAILABLE_CHECK,
@@ -22,8 +24,6 @@ from .const import (
 )
 from .coordinator import GreeConfigEntry, GreeCoordinator
 from .entity import GreeEntity, GreeEntityDescription
-from .gree_api import GreeProp, TemperatureUnits
-from .gree_device import GreeDevice
 
 _LOGGER = logging.getLogger(__name__)
